Issue #599 Moving get_asset_group_queries from data/services to data/queries
